Version final a entregar

+++ src/app/dashboard/pages/teachers/teachers.component.ts
@@ -2,11 +2,11 @@ import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { TeachersActions } from './store/teachers.actions';
 import { Observable } from 'rxjs';
-import { TeacherWithCourse } from './models';
 import { selectTeachers} from './store/teachers.selectors';
 import { MatDialog } from '@angular/material/dialog';
 import { TeacherFormDialogComponent } from './components/teacher-form-dialog/teacher-form-dialog.component';
 import { selectIsAdmin } from 'src/app/store/auth/auth.selectors';
+import { Teacher } from './models';
 
 @Component({
   selector: 'app-teachers',
@@ -16,7 +16,7 @@ import { selectIsAdmin } from 'src/app/store/auth/auth.selectors';
 
 export class TeachersComponent implements OnInit {
 
-  teachers$: Observable<TeacherWithCourse[]>;
+  teachers$: Observable<Teacher[]>;
 
   isAdmin$ : Observable<boolean>;
 
@@ -33,7 +33,7 @@ export class TeachersComponent implements OnInit {
 
   
 
-  public displayedColumns = ['id','nombre','email','curso','acciones'];
+  public displayedColumns = ['id','nombre','email','acciones'];
 
 
 onCreateTeacher(): void
@@ -44,6 +44,9 @@ onCreateTeacher(): void
 
 onDeleteTeacher(teacher: number): void 
 {
-  this.store.dispatch(TeachersActions.deleteTeacher({ id: teacher}));
+  if(confirm(`¿Está seguro que desea eliminar el profesor?`))
+  {
+    this.store.dispatch(TeachersActions.deleteTeacher({ id: teacher}));
+  }
 }
 }
